[pdf.js] Update to release 1.5.188
**Extern:** I updated the extern with generator.

A few  manual adjustments to re-order entries so the diff isn't so scary,
and to correct workerSrc attribute which the generator mistakenly classed as a fn.

This also removes a number of externs. These weren't exposed in 1.1.3 nor are they
in current 1.5.188, so they seem to have been mistakenly added.
The then/cancel/getPage/getViewport properties are to the promise returned from
.getDocument; not from pdfjs per se.

AnnotationUtils/DefaultTextLayerFactory seems to come from viewer.js,
not exported in the dist pdf.js.