Remove test and demo files from libphonenumber
Some of those files have the same goog.provide and define the same symbols
as the "main" files, creating errors due to the wrong version of some
symbols being loaded.

One example is metadatafortesting.js which defines a country to metadata
mapping for a reduced set of countries. When this mapping ends being
loaded instead of the main one, user of this library might get errors like
"invalid country calling code" for valid country codes. This reduced
country set metadata seems useful for testing libphonenumber as it must
load faster, but testing libphonenumber is not the intended usage of the
cljsjs package.

I'm not sure what determines which version of the duplicate symbols is the
one that ends being used, it seems to be that it depends on a
non-deterministic loading error during compilation at some part of the
google closure compiler. What's clear to me is that having two files
to define the same symbol with different meanings under the same
goog.provide is ambiguous at the very least.