[tock] Add tock, an accurate timer library
Tock is a self-correcting, accurate JavaScript timer. Typically JavaScript
timing can be done with setInterval and setTimeout, but they don't have any
guarantees that e.g. a timeout of 1 second will be exactly 1 second.

Tock solves this problem by keeping track of the delay between the expected and
actual timeout resolution times, and adjusts its timeout to account for this
delay.