Update patch version of mixpanel to restore "people" in externs
The most recent mixpanel update used the raw
javascript-externs-generator output, but this does not include some of
the externs that are needed, namely the "people" api functions.

Add those back, and updated the version, externs note, and even added
a note to the README to hopefully make next time easier.