[react-swipeable-views] Add package
Currently builds off a commit from my branch that adds support
for building as a UMD, because the main project builds as an es6
module. I have a PR up to merge the branch in, but the project
owner has been focusing on other projects and has not been very
responsive.