[highcharts] Support 8.1.0 version of highcharts package
Updated externs with automatically generated code from the same library it was
used before.

Update library version number and reset patch index in boot file.

Update checksums after generating the package with boot.

Additional manual entries in externs file are not included yet as it make
sense to be included in separated commit for legibility.