[p5] Update p5 to 1.7.0 release
Updates cljsjs/p5 from 1.4.0-0 to 1.7.0-0 to match the upstream release
https://github.com/processing/p5.js/releases/tag/v1.7.0.

Externs file was amended to include any additional entries detected by the jmmk
javascripte externs generator, but ignored any removals. It appears much of this
library is dynamically generated to add it's own custom type checking errors and
the jmmk generator does not detect much of the dynamically generated classes.

I'm in the process of updating the https://github.com/quil/quil library to use
the latest release of p5 and will amend the externs file if I find any
additional externs are missing. In the future we may move this dependency into
the quil.jar directly as I know cljsjs is deprecated, but for now I'm hoping to
keep the existing release structure working with cljsjs providing the p5
dependency.